Add support for FFmpeg as a decoder. Many breaking changes. (#47)
* Add support for FFmpeg as a decoding backend

* Refactor Symphonia and FFmpeg to match

* Finally made both Symphonia and FFmpeg APIs compatible

* Adding benchmarks and API changes

* Allow batch decoding directly to NumPy arrays

* Fix unit tests

* Handle file-not-found error in FFmpeg decoder

* dockerignore our `audio-for-tests` directory

* Actions: Fix typo

* Makefile: avoid unnecessary whitespace

* Adjust file-not-found error handler in FFmpeg decoder

* Actions: Compile Rust unit tests with FFmpeg

* Start writing Python benchmarks

* actions: Install yasm so we can build ffmpeg

* Exclude even more stuff from dockerignore

+++ src/backend/common.rs
@@ -1,3 +1,3 @@
-pub fn milliseconds_to_frames(frame_rate_hz: u32, duration_milliseconds: u64) -> u64 {
-    (duration_milliseconds * frame_rate_hz as u64) / 1000
+pub fn milliseconds_to_frames(frame_rate_hz: u32, duration_milliseconds: usize) -> usize {
+    (duration_milliseconds * frame_rate_hz as usize) / 1000
 }
