+++ src/backend/resample/babycat_lanczos.rs
@@ -29,14 +29,14 @@ fn compute_sample(
     channel_idx: usize,
     num_channels: usize,
 ) -> f32 {
-    let num_input_frames: u64 = input_audio.len() as u64 / num_channels as u64;
+    let num_input_frames: usize = input_audio.len() / num_channels as usize;
     let a: f32 = KERNEL_A as f32;
     let x_floor = frame_idx as i64;
     let i_start = x_floor - a as i64 + 1;
     let i_end = x_floor + a as i64 + 1;
     let mut the_sample: f32 = 0.0_f32;
     for i in i_start..i_end {
-        if (i as u64) < num_input_frames {
+        if (i as usize) < num_input_frames {
             the_sample += get(input_audio, i as usize, channel_idx, num_channels)
                 * lanczos_kernel(frame_idx - i as f32, a)
         }
@@ -47,7 +47,7 @@ fn compute_sample(
 pub fn resample(
     input_frame_rate_hz: u32,
     output_frame_rate_hz: u32,
-    num_channels: u32,
+    num_channels: u16,
     input_audio: &[f32],
 ) -> Result<Vec<f32>, Error> {
     validate_args(input_frame_rate_hz, output_frame_rate_hz, num_channels)?;
